refactor(util): 移动并重命名 eventBus.js
1.采用 kebab 风格重命名
2.移动至 core/utils 目录
